Move NotificationScreen back and rework it in the manner of TranslatableRichTextViewer.
